\name{cqi_regex2id}
\alias{cqi_regex2id}

\title{* cqi_regex2id *}

\description{
TODO...
}

\usage{
	cqi_regex2id(attribute, regex)
}

\arguments{
	\item{attribute}{(string) the name of a positional attribute.}
	\item{regex}{(string) a regular expression.}
}

\details{
TODO...
}

\value{
TODO...
}

\author{
	Bernard Desgraupes - \email{bernard.desgraupes@u-paris10.fr} - University Paris-10.\cr
	Sylvain Loiseau - \email{sylvain.loiseau@univ-paris13.fr} - University Paris-13.\cr
}

\source{
	The IMS Open Corpus Workbench (CWB) \url{http://cwb.sourceforge.net/}
}

\references{
	\url{http://cwb.sourceforge.net/documentation.php}
}

\seealso{
TODO...
}

\examples{
# TODO...
}

\keyword{corpus position}