\name{cqi_struc2cpos}
\alias{cqi_struc2cpos}
\alias{cqi_struc2str}

\title{Convert structural indices}

\description{
TODO...
}

\usage{
	cqi_struc2cpos(attribute, struc)
	cqi_struc2str(attribute, ids)
}

\arguments{
	\item{attribute}{(string) the qualified name of an attribute.}
	\item{struc}{(integer) }
	\item{ids}{(integer vector)}
}

\details{
TODO...
}

\value{
TODO...
}

\author{
	Bernard Desgraupes - \email{bernard.desgraupes@u-paris10.fr} - University Paris-10.\cr
	Sylvain Loiseau - \email{sylvain.loiseau@univ-paris13.fr} - University Paris-13.\cr
}

\source{
	The IMS Open Corpus Workbench (CWB) \url{http://cwb.sourceforge.net/}
}

\references{
	\url{http://cwb.sourceforge.net/documentation.php}
}

\seealso{
\code{\link{cqi_alg2cpos}},
\code{\link{cqi_cpos2alg}},
\code{\link{cqi_cpos2id}},
\code{\link{cqi_cpos2lbound}},
\code{\link{cqi_cpos2rbound}},
\code{\link{cqi_cpos2str}},
\code{\link{cqi_cpos2struc}},
\code{\link{cqi_id2cpos}},
\code{\link{cqi_id2freq}},
\code{\link{cqi_id2str}},
\code{\link{cqi_regex2id}},
\code{\link{cqi_str2id}}
}

\examples{
# TODO...
}

\keyword{corpus position}
