\name{cqp_undump}
\alias{cqp_undump}

\title{* cqp_undump *}

\description{
TODO...
}

\usage{
	cqp_undump(file, corpus, subcorpus, target=FALSE, keyword=FALSE)
}

\arguments{
	\item{file}{(\emph{string}) TODO...}
	\item{subcorpus}{(\emph{string}) qualified name of a subcorpus.}
	\item{corpus}{(\emph{string}) the name of a corpus (in uppercase letters).}
	\item{target}{TODO...}
	\item{keyword}{TODO...}
}

\details{
TODO...
}

\value{
\code{cqp_undump} returns 
}

\author{
	Bernard Desgraupes - \email{bernard.desgraupes@u-paris10.fr} - University Paris-10.\cr
	Sylvain Loiseau - \email{sylvain.loiseau@univ-paris13.fr} - University Paris-13.\cr
}

\source{
	The IMS Open Corpus Workbench (CWB) at \url{http://cwb.sourceforge.net/}
}

\references{
	\url{http://cwb.sourceforge.net/documentation.php}
}

\seealso{
TODO...
}

\examples{
\dontrun{
}
}

