\name{rcqp}
\alias{rcqp}
\alias{rcqp-package}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{~ Corpus Query Protocol ~}
\description{
  \tabular{ll}{
    Package: \tab rcqp\cr
    Type: \tab Package\cr
    Version: \tab 0.1\cr
    Date: \tab 2012-01-20\cr
    License: \tab  GPL (>= 2) \cr
  }
}

\details{
  \code{rcqp} is an implementation in R of the Corpus Query Protocol.
}

\section{Author}{
	Bernard Desgraupes - \email{bernard.desgraupes@u-paris10.fr} - University Paris-10.\cr
	Sylvain Loiseau - \email{sylvain.loiseau@univ-paris13.fr} - University Paris-13.\cr
}

\references{
\itemize{
	\item \emph{CQP Query Language Tutorial}: \url{http://cwb.sourceforge.net/files/CQP_Tutorial.pdf}

	\item \emph{Corpus Encoding Tutorial}: \url{http://cwb.sourceforge.net/files/CWB_Encoding_Tutorial.pdf}

}
}

\seealso{
\code{\link{cqi_query}}
}

\keyword{package}
